Runs DestroyNotify hook on WM_DELETE_WINDOW
Add a condition for the mlx loop to use, if set, the DestroyNotify
hook when receiving a WM_DELETE_WINDOW ClientMessage event.

This does not affect the ClientMessage hook and will not break previous
code. This is to avoid using the ClientMessage hook as a way to close
the application as it is not guaranteed the Clientmessage will be a
WM_DELETE_WINDOW.

It also streamlines the mlx to work similarily to its MacOS counterpart.